refactor: entry name
