+++ packages/rax-text/demo/index/index.server.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import { createElement } from 'rax';
 import { renderToString } from 'rax-server-renderer';
-import App from './index';
+import App from './app';
 
 export default function() {
   return renderToString(<App />);
